requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Flask-WTF==0.9.5
gunicorn==18.0
python-decouple==2.3
Unipath==1.0
